Core functionality now working, mostly
Another step towards getting this working. Now you can pass in a .out file
and it will parse out the unique files, count the lines of the functions
found within those files, and finally output the stats for it. It needs more
work before it can start printing metrics, but that is the next step.