Trying to get one line functions working
This was failing when it encountered a series of functions defined on only one line.
Obviously this was an edge case I hadn't thought about and so now I have to fix it.
There is a bunch of debugging statements in here so I'm splitting this off into it's
own branch until I get this feature working.